sixteenth homework

+++ app/src/main/java/com/example/playlistmaker/data/impl/PlaylistRepositoryImpl.kt
@@ -8,13 +8,18 @@ import com.example.playlistmaker.domain.db.PlaylistRepository
 import com.example.playlistmaker.domain.models.Playlist
 import com.example.playlistmaker.domain.models.Track
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.flow
 
 class PlaylistRepositoryImpl(
     private val appDatabase: AppDatabase,
     private val playlistDbConverter: PlaylistDbConverter
 ) : PlaylistRepository {
+    override fun getPlaylistById(id: Int): Flow<Playlist> = flow {
+        appDatabase.playlistDao().getPlaylistById(id).collect {
+            emit(playlistDbConverter.map(it))
+        }
+    }
+
     override fun getPlaylists(): Flow<List<Playlist>> = flow {
         appDatabase.playlistDao().getPlaylists().collect {
             val playlists = convertFromEntity(it)
@@ -24,6 +29,12 @@ class PlaylistRepositoryImpl(
 
     override suspend fun savePlaylist(playlist: Playlist) {
         appDatabase.playlistDao().savePlaylist(playlistDbConverter.map(playlist))
+    }
+
+    override suspend fun checkTrackId(id: Int): Boolean {
+        val tracksId: List<List<Int>> = appDatabase.playlistDao().getPlaylistsTracksId().map { playlistDbConverter.map(it) }
+        tracksId.map { if (it.contains(id)) return true }
+        return false
     }
 
     override suspend fun updatePlaylist(playlist: Playlist) {
@@ -42,8 +53,17 @@ class PlaylistRepositoryImpl(
         appDatabase.playlistTrackDao().saveTrack(playlistDbConverter.map(track))
     }
 
-    override suspend fun deleteTrack(track: Track) {
-        appDatabase.playlistTrackDao().deleteTrack(playlistDbConverter.map(track))
+    override suspend fun deleteTrack(track: Track, playlist: Playlist): Flow<Playlist> = flow {
+        (playlist.trackList as ArrayList<Int>).remove(track.trackId)
+        playlist.trackCountInt = playlist.trackList.size
+        updatePlaylist(playlist)
+        if (checkTrackId(track.trackId!!))
+            appDatabase.playlistTrackDao().deleteTrack(playlistDbConverter.map(track))
+        emit(playlist)
+    }
+
+    override suspend fun deletePlaylist(playlist: Playlist) {
+        appDatabase.playlistDao().deletePlaylist(playlistDbConverter.map(playlist))
     }
 
     private fun convertFromTrackEntity(tracks: List<PlaylistTrackEntity>) : List<Track> {
